# [Abandoned] Fix number of lines reported by function_body_length and type_body_length