--- a/.swiftlint.yml
+++ b/.swiftlint.yml
@@ -44,6 +44,10 @@ identifier_name:
 line_length: 120
 number_separator:
   minimum_length: 5
+type_body_length:
+    - 250
+function_body_length:
+    - 50
 
 custom_rules:
   rule_id: