--- a/Rules.md
+++ b/Rules.md
@@ -11920,7 +11920,6 @@ let abc = 0
 let abc = 0
 let abc = 0
 }
-
 ```
 
 ```swift
@@ -12127,7 +12126,6 @@ class Abc {
 
 
 }
-
 ```
 
 ```swift
@@ -12334,7 +12332,6 @@ class Abc {
 // this is a comment
 // this is a comment
 }
-
 ```
 
 ```swift
@@ -12543,7 +12540,6 @@ let abc = 0
 a multiline comment
 */
 }
-
 ```
 
 ```swift
@@ -12748,7 +12744,6 @@ let abc = 0
 let abc = 0
 let abc = 0
 }
-
 ```
 
 ```swift
@@ -12955,7 +12950,6 @@ struct Abc {
 
 
 }
-
 ```
 
 ```swift
@@ -13162,7 +13156,6 @@ struct Abc {
 // this is a comment
 // this is a comment
 }
-
 ```
 
 ```swift
@@ -13371,7 +13364,6 @@ let abc = 0
 a multiline comment
 */
 }
-
 ```
 
 ```swift
@@ -13576,7 +13568,6 @@ let abc = 0
 let abc = 0
 let abc = 0
 }
-
 ```
 
 ```swift
@@ -13783,7 +13774,6 @@ enum Abc {
 
 
 }
-
 ```
 
 ```swift
@@ -13990,7 +13980,6 @@ enum Abc {
 // this is a comment
 // this is a comment
 }
-
 ```
 
 ```swift
@@ -14199,7 +14188,6 @@ let abc = 0
 a multiline comment
 */
 }
-
 ```
 
 </details>
@@ -14410,7 +14398,6 @@ let abc = 0
 let abc = 0
 let abc = 0
 }
-
 ```
 
 ```swift
@@ -14617,7 +14604,6 @@ let abc = 0
 let abc = 0
 let abc = 0
 }
-
 ```
 
 ```swift
@@ -14824,7 +14810,6 @@ let abc = 0
 let abc = 0
 let abc = 0
 }
-
 ```
 
 </details>