--- a/Source/SwiftLintFramework/Rules/QuickDiscouragedCallRuleExamples.swift
+++ b/Source/SwiftLintFramework/Rules/QuickDiscouragedCallRuleExamples.swift
@@ -8,8 +8,6 @@
 
 import Foundation
 
-// swiftlint:disable type_body_length
-
 internal struct QuickDiscouragedCallRuleExamples {
     static let nonTriggeringExamples: [String] = [
         "class TotoTests: QuickSpec {\n" +