This commit implements a fix for the issue described in #1750. The following rules are returning the incorrect number of lines:

* `function_body_length`
* `type_body_length`

This PR contains several commits to make the code review easier.