@marcelofabri 2017-08-07
Reminder to my future self: run `oss-check` locally to see if this impacts the performance.