@ornithocoder 2017-08-07
@marcelofabri I got this:

```
personal-fork-swiftlint $ ./script/oss-check --branch blank_line --iteration 10
Preparing osscheck/builds (identifier builds)
Previous HEAD position was 81fb8a64... Add test for case which was reporting incorrect number of lines
HEAD is now at ef67adca... rename RulesDocsOptions to GenerateDocsOptions
Message: Linting Aerial with this PR took 0.3s vs 0.31s on master (3% faster)
Message: Linting Alamofire with this PR took 2.25s vs 2.34s on master (3% faster)
Message: Linting Firefox with this PR took 9.57s vs 10.55s on master (9% faster)
Message: Linting Kickstarter with this PR took 13.84s vs 13.91s on master (0% faster)
Message: Linting Moya with this PR took 0.97s vs 0.63s on master (53% slower)
Message: Linting Nimble with this PR took 1.2s vs 1.25s on master (4% faster)
Message: Linting Quick with this PR took 0.4s vs 0.42s on master (4% faster)
Message: Linting Realm with this PR took 2.08s vs 2.02s on master (2% slower)
Message: Linting SourceKitten with this PR took 0.75s vs 0.72s on master (4% slower)
Message: Linting Sourcery with this PR took 3.59s vs 3.75s on master (4% faster)
Message: Linting Swift with this PR took 11.08s vs 11.13s on master (0% faster)
Message: Linting WordPress with this PR took 9.42s vs 9.72s on master (3% faster)
```