@ornithocoder 2017-08-07
@marcelofabri I'll try to cache the result. Changing it to WIP for now.