@ornithocoder 2017-08-07
@marcelofabri I've tried cache, but got the same results. Cache implementation can be found in this gist https://gist.github.com/ornithocoder/eb4b407fdbd8e74cbeb3d7bdb0254bd9.

The result is pretty much the same, but honestly, I prefer the cache implementation better and am thinking of pushing it. What do you think?