@marcelofabri 2017-08-07
Since the results are the same, I wouldn't add the cache to keep things simpler. One thing that *might* help is to only evaluate `let isEmpty = line.content.trimmingCharacters(in: .whitespaces).isEmpty` if `comment` is `true`.