@ornithocoder 2017-08-09
@marcelofabri I decided to give it another try and increase the number of iterations to 20. These are the results:

**Without Caching**

```
personal-fork-swiftlint $ ./script/oss-check --branch blank_line --iteration 20
Cloning Aerial
Cloning Alamofire
Cloning Firefox
Cloning Kickstarter
Cloning Moya
Cloning Nimble
Cloning Quick
Cloning Realm
Cloning SourceKitten
Cloning Sourcery
Cloning Swift
Cloning WordPress
Building branch
Preparing osscheck/builds (identifier builds)
Linting 20 iterations of Aerial with branch: 1..2..3..4..5..6..7..8..9..10..11..12..13..14..15..16..17..18..19..20
Linting 20 iterations of Alamofire with branch: 1..2..3..4..5..6..7..8..9..10..11..12..13..14..15..16..17..18..19..20
Linting 20 iterations of Firefox with branch: 1..2..3..4..5..6..7..8..9..10..11..12..13..14..15..16..17..18..19..20
Linting 20 iterations of Kickstarter with branch: 1..2..3..4..5..6..7..8..9..10..11..12..13..14..15..16..17..18..19..20
Linting 20 iterations of Moya with branch: 1..2..3..4..5..6..7..8..9..10..11..12..13..14..15..16..17..18..19..20
Linting 20 iterations of Nimble with branch: 1..2..3..4..5..6..7..8..9..10..11..12..13..14..15..16..17..18..19..20
Linting 20 iterations of Quick with branch: 1..2..3..4..5..6..7..8..9..10..11..12..13..14..15..16..17..18..19..20
Linting 20 iterations of Realm with branch: 1..2..3..4..5..6..7..8..9..10..11..12..13..14..15..16..17..18..19..20
Linting 20 iterations of SourceKitten with branch: 1..2..3..4..5..6..7..8..9..10..11..12..13..14..15..16..17..18..19..20
Linting 20 iterations of Sourcery with branch: 1..2..3..4..5..6..7..8..9..10..11..12..13..14..15..16..17..18..19..20
Linting 20 iterations of Swift with branch: 1..2..3..4..5..6..7..8..9..10..11..12..13..14..15..16..17..18..19..20
Linting 20 iterations of WordPress with branch: 1..2..3..4..5..6..7..8..9..10..11..12..13..14..15..16..17..18..19..20
Building master
Previous HEAD position was 9959280a... Add .testFunctionBodyLengthsWithIfs() to Linux tests
HEAD is now at ef67adca... rename RulesDocsOptions to GenerateDocsOptions
Linting 20 iterations of Aerial with master: 1..2..3..4..5..6..7..8..9..10..11..12..13..14..15..16..17..18..19..20
Linting 20 iterations of Alamofire with master: 1..2..3..4..5..6..7..8..9..10..11..12..13..14..15..16..17..18..19..20
Linting 20 iterations of Firefox with master: 1..2..3..4..5..6..7..8..9..10..11..12..13..14..15..16..17..18..19..20
Linting 20 iterations of Kickstarter with master: 1..2..3..4..5..6..7..8..9..10..11..12..13..14..15..16..17..18..19..20
Linting 20 iterations of Moya with master: 1..2..3..4..5..6..7..8..9..10..11..12..13..14..15..16..17..18..19..20
Linting 20 iterations of Nimble with master: 1..2..3..4..5..6..7..8..9..10..11..12..13..14..15..16..17..18..19..20
Linting 20 iterations of Quick with master: 1..2..3..4..5..6..7..8..9..10..11..12..13..14..15..16..17..18..19..20
Linting 20 iterations of Realm with master: 1..2..3..4..5..6..7..8..9..10..11..12..13..14..15..16..17..18..19..20
Linting 20 iterations of SourceKitten with master: 1..2..3..4..5..6..7..8..9..10..11..12..13..14..15..16..17..18..19..20
Linting 20 iterations of Sourcery with master: 1..2..3..4..5..6..7..8..9..10..11..12..13..14..15..16..17..18..19..20
Linting 20 iterations of Swift with master: 1..2..3..4..5..6..7..8..9..10..11..12..13..14..15..16..17..18..19..20
Linting 20 iterations of WordPress with master: 1..2..3..4..5..6..7..8..9..10..11..12..13..14..15..16..17..18..19..20
Message: Linting Aerial with this PR took 0.31s vs 0.31s on master (0% slower)
Message: Linting Alamofire with this PR took 2.43s vs 2.39s on master (1% slower)
Message: Linting Firefox with this PR took 10.49s vs 10.12s on master (3% slower)
Message: Linting Kickstarter with this PR took 15.24s vs 15.44s on master (1% faster)
Message: Linting Moya with this PR took 1.02s vs 0.7s on master (45% slower)
Message: Linting Nimble with this PR took 1.4s vs 1.32s on master (6% slower)
Message: Linting Quick with this PR took 0.42s vs 0.44s on master (4% faster)
Message: Linting Realm with this PR took 2.23s vs 2.38s on master (6% faster)
Message: Linting SourceKitten with this PR took 0.8s vs 0.78s on master (2% slower)
Message: Linting Sourcery with this PR took 3.94s vs 3.86s on master (2% slower)
Message: Linting Swift with this PR took 11.08s vs 11.01s on master (0% slower)
Message: Linting WordPress with this PR took 10.0s vs 10.04s on master (0% faster)
```

**With Caching**

```
$ ./script/oss-check --branch blank_line --iteration 20
Cloning Aerial
Cloning Alamofire
Cloning Firefox
Cloning Kickstarter
Cloning Moya
Cloning Nimble
Cloning Quick
Cloning Realm
Cloning SourceKitten
Cloning Sourcery
Cloning Swift
Cloning WordPress
Building branch
Preparing osscheck/builds (identifier builds)
Linting 20 iterations of Aerial with branch: 1..2..3..4..5..6..7..8..9..10..11..12..13..14..15..16..17..18..19..20
Linting 20 iterations of Alamofire with branch: 1..2..3..4..5..6..7..8..9..10..11..12..13..14..15..16..17..18..19..20
Linting 20 iterations of Firefox with branch: 1..2..3..4..5..6..7..8..9..10..11..12..13..14..15..16..17..18..19..20
Linting 20 iterations of Kickstarter with branch: 1..2..3..4..5..6..7..8..9..10..11..12..13..14..15..16..17..18..19..20
Linting 20 iterations of Moya with branch: 1..2..3..4..5..6..7..8..9..10..11..12..13..14..15..16..17..18..19..20
Linting 20 iterations of Nimble with branch: 1..2..3..4..5..6..7..8..9..10..11..12..13..14..15..16..17..18..19..20
Linting 20 iterations of Quick with branch: 1..2..3..4..5..6..7..8..9..10..11..12..13..14..15..16..17..18..19..20
Linting 20 iterations of Realm with branch: 1..2..3..4..5..6..7..8..9..10..11..12..13..14..15..16..17..18..19..20
Linting 20 iterations of SourceKitten with branch: 1..2..3..4..5..6..7..8..9..10..11..12..13..14..15..16..17..18..19..20
Linting 20 iterations of Sourcery with branch: 1..2..3..4..5..6..7..8..9..10..11..12..13..14..15..16..17..18..19..20
Linting 20 iterations of Swift with branch: 1..2..3..4..5..6..7..8..9..10..11..12..13..14..15..16..17..18..19..20
Linting 20 iterations of WordPress with branch: 1..2..3..4..5..6..7..8..9..10..11..12..13..14..15..16..17..18..19..20
Building master
Previous HEAD position was 369c4f18... Cache again
HEAD is now at ef67adca... rename RulesDocsOptions to GenerateDocsOptions
Linting 20 iterations of Aerial with master: 1..2..3..4..5..6..7..8..9..10..11..12..13..14..15..16..17..18..19..20
Linting 20 iterations of Alamofire with master: 1..2..3..4..5..6..7..8..9..10..11..12..13..14..15..16..17..18..19..20
Linting 20 iterations of Firefox with master: 1..2..3..4..5..6..7..8..9..10..11..12..13..14..15..16..17..18..19..20
Linting 20 iterations of Kickstarter with master: 1..2..3..4..5..6..7..8..9..10..11..12..13..14..15..16..17..18..19..20
Linting 20 iterations of Moya with master: 1..2..3..4..5..6..7..8..9..10..11..12..13..14..15..16..17..18..19..20
Linting 20 iterations of Nimble with master: 1..2..3..4..5..6..7..8..9..10..11..12..13..14..15..16..17..18..19..20
Linting 20 iterations of Quick with master: 1..2..3..4..5..6..7..8..9..10..11..12..13..14..15..16..17..18..19..20
Linting 20 iterations of Realm with master: 1..2..3..4..5..6..7..8..9..10..11..12..13..14..15..16..17..18..19..20
Linting 20 iterations of SourceKitten with master: 1..2..3..4..5..6..7..8..9..10..11..12..13..14..15..16..17..18..19..20
Linting 20 iterations of Sourcery with master: 1..2..3..4..5..6..7..8..9..10..11..12..13..14..15..16..17..18..19..20
Linting 20 iterations of Swift with master: 1..2..3..4..5..6..7..8..9..10..11..12..13..14..15..16..17..18..19..20
Linting 20 iterations of WordPress with master: 1..2..3..4..5..6..7..8..9..10..11..12..13..14..15..16..17..18..19..20
Message: Linting Aerial with this PR took 0.33s vs 0.32s on master (3% slower)
Message: Linting Alamofire with this PR took 1.44s vs 1.57s on master (8% faster)
Message: Linting Firefox with this PR took 10.37s vs 11.21s on master (7% faster)
Message: Linting Kickstarter with this PR took 15.48s vs 16.28s on master (4% faster)
Message: Linting Moya with this PR took 0.99s vs 0.83s on master (19% slower)
Message: Linting Nimble with this PR took 1.29s vs 1.45s on master (11% faster)
Message: Linting Quick with this PR took 0.49s vs 0.46s on master (6% slower)
Message: Linting Realm with this PR took 2.19s vs 2.36s on master (7% faster)
Message: Linting SourceKitten with this PR took 0.75s vs 0.91s on master (17% faster)
Message: Linting Sourcery with this PR took 3.83s vs 3.94s on master (2% faster)
Message: Linting Swift with this PR took 11.03s vs 11.29s on master (2% faster)
Message: Linting WordPress with this PR took 9.95s vs 9.55s on master (4% slower)
```